back/requirements.txt
# Django and Django REST Framework
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0

# Database
psycopg2-binary==2.9.10

# CORS handling
django-cors-headers==4.6.0

# Caching (Redis)
django-redis==5.4.0
redis==5.2.1

# Image processing and QR codes
Pillow==11.0.0
qrcode[pil]==8.0

# Environment variables
python-decouple==3.8

# HTTP requests
requests==2.32.3

# Security and validation
django-ratelimit==4.1.0

# Development dependencies
django-debug-toolbar==4.4.6

# Testing
coverage==7.6.10
factory-boy==3.3.1

# Logging
structlog==24.4.0

# Time zone support
pytz==2024.2

# Production server
gunicorn==23.0.0
whitenoise==6.8.2

# Monitoring
sentry-sdk==2.19.2

# Analytics and Visualization
plotly==5.22.0
pandas==2.3.1
numpy==2.1.3